+++ CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 
-- Bucket migration for packages-dev.wazuh.com ([#1712](https://github.com/wazuh/wazuh-ansible/pull/1712))- None
+- Bucket migration for packages-dev.wazuh.info ([#1712](https://github.com/wazuh/wazuh-ansible/pull/1712))- None
 
 ### Fixed